feat#1 :: @Trace 삭제, 예외 처리 및 테스트 코드 수정

+++ src/main/java/solvit/teachmon/domain/management/teacher/application/facade/ManagementTeacherFacadeService.java
@@ -10,7 +10,7 @@ import solvit.teachmon.domain.management.teacher.presentation.dto.response.Teach
 import solvit.teachmon.domain.supervision.domain.repository.SupervisionScheduleRepository;
 import solvit.teachmon.domain.user.domain.entity.TeacherEntity;
 import solvit.teachmon.domain.user.domain.repository.TeacherRepository;
-import solvit.teachmon.global.annotation.Trace;
+import solvit.teachmon.domain.user.exception.TeacherNotFoundException;
 import solvit.teachmon.global.enums.WeekDay;
 
 import java.util.ArrayList;
@@ -23,32 +23,32 @@ public class ManagementTeacherFacadeService {
     private final SupervisionScheduleRepository supervisionScheduleRepository;
     private final SupervisionBanDayRepository supervisionBanDayRepository;
 
-    @Trace
+    @Transactional(readOnly = true)
     public List<TeacherListResponse> getAllTeachers(String query) {
         return supervisionScheduleRepository.countTeacherSupervision(query);
     }
 
-    @Trace
     @Transactional
     public void updateTeacher(TeacherUpdateRequest updateRequest, Long teacherId) {
         TeacherEntity teacher = teacherRepository.findById(teacherId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 ID의 교사를 찾을 수 없습니다"));
+                .orElseThrow(TeacherNotFoundException::new);
 
         teacher.changeRole(updateRequest.role());
         teacher.changeName(updateRequest.name());
     }
 
-    @Trace
     @Transactional
     public void deleteTeacher(Long teacherId) {
+        if (!teacherRepository.existsById(teacherId)) {
+            throw new TeacherNotFoundException();
+        }
         teacherRepository.deleteById(teacherId);
     }
 
-    @Trace
     @Transactional
     public void setTeacherBanDay(Long teacherId, List<WeekDay> banDays) {
         TeacherEntity teacher = teacherRepository.findById(teacherId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 ID의 교사를 찾을 수 없습니다"));
+                .orElseThrow(TeacherNotFoundException::new);
 
         supervisionBanDayRepository.deleteAllByTeacherId(teacherId);
